requirements.txt
python-dotenv==1.0.0

gspread~=5.8.0
oauth2client~=4.1.3
google-api-python-client~=2.86.0
prettytable~=3.7.0

bs4~=0.0.1
beautifulsoup4~=4.12.2
langchain~=0.0.166
pydantic~=1.10.6
psutil~=5.9.5